# docs/openapi.yaml
openapi: 3.0.3

info:
  title: Distributed Job Queue API
  version: 1.0.0
  description: |
    REST API for a priority-based distributed job queue. Jobs are persisted to PostgreSQL, dispatched to RabbitMQ quorum queues, and processed by worker instances.

    **Envelope**: All responses use `CommonResponse`. On errors, `data` is omitted.
    **Tracing**: Responses may include `traceId`/`spanId` when W3C Trace Context headers are present.
    **Rate limiting**: Enforced at the NGINX Ingress layer, not within the application.
  contact:
    name: nadimnesar
  license:
    name: MIT

externalDocs:
  description: Project README with setup instructions and architecture details
  url: https://github.com/nadimnesar/distributed-job-queue#readme

servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://192.168.49.2
    description: Kubernetes via NGINX Ingress

tags:
  - name: Jobs
    description: Job lifecycle — create, list, filter, cancel, revive.
  - name: Dashboard
    description: Operational metrics — job summary and queue depths.

paths:
  /api/v1/job/create:
    post:
      tags: [ Jobs ]
      operationId: createJob
      summary: Create a new job
      description: |
        Validates the request body (DAG dependency rules, priority hierarchy),
        persists the job, and publishes to RabbitMQ after transaction commit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
            examples:
              simple:
                summary: Simple job, no dependencies
                value:
                  priority: HIGH
                  type: EMAIL_SENDING
                  payload: Send welcome email to user 123
                  maxAttemptCount: 3
              withDependencies:
                summary: Job with dependencies
                value:
                  priority: HIGH
                  type: PAYMENT_SENDING
                  payload: Process refund for order 456
                  dependencies:
                    - 550e8400-e29b-41d4-a716-446655440000
                    - 6ba7b810-9dad-11d1-80b4-00c04fd430c8
                  maxAttemptCount: 5
              minimal:
                summary: Minimal payload (optional fields omitted)
                value:
                  priority: MEDIUM
                  type: EMAIL_SENDING
                  payload: Send notification
      responses:
        '201':
          description: Job created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseJobResponse'
              example:
                message: Job is created successfully.
                code: 201
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  priority: HIGH
                  type: EMAIL_SENDING
                  status: PENDING
                  dependents: [ ]
                  dependencies: [ ]
                  attemptCount: 0
                  maxAttemptCount: 3
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/jobs:
    get:
      tags: [ Jobs ]
      operationId: getJobs
      summary: List all jobs (paginated)
      description: Returns a page of all jobs, newest first. Both `page` and `size` are optional with defaults.
      parameters:
        - name: page
          in: query
          required: false
          description: Zero-indexed page number. Defaults to 0.
          schema:
            type: integer
            minimum: 0
            default: 0
          examples:
            first:
              summary: First page
              value: 0
            second:
              summary: Second page
              value: 1
        - name: size
          in: query
          required: false
          description: Number of jobs per page. Defaults to 20, must be >= 1.
          schema:
            type: integer
            minimum: 1
            default: 20
          examples:
            default:
              summary: Default page size (20)
              value: 20
            large:
              summary: Large page
              value: 200
      responses:
        '200':
          description: A page of jobs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseJobResponseList'
              example:
                message: Operation Successful.
                code: 200
                data:
                  - id: 550e8400-e29b-41d4-a716-446655440000
                    priority: HIGH
                    type: EMAIL_SENDING
                    status: PENDING
                    dependents: [ ]
                    dependencies: [ ]
                    attemptCount: 0
                    maxAttemptCount: 5
                  - id: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
                    priority: MEDIUM
                    type: PAYMENT_SENDING
                    status: COMPLETED
                    dependents: [ ]
                    dependencies: [ ]
                    attemptCount: 1
                    maxAttemptCount: 5
                    result: Payment processed successfully
                    startedAt: '2026-06-25T10:15:30Z'
                    completedAt: '2026-06-25T10:15:32Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/jobs/{id}:
    get:
      tags: [ Jobs ]
      operationId: getJobById
      summary: Get a job by ID
      description: Returns the full job representation. Null fields are omitted.
      parameters:
        - $ref: '#/components/parameters/JobIdPath'
      responses:
        '200':
          description: The requested job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseJobResponse'
              example:
                message: Operation Successful.
                code: 200
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  priority: HIGH
                  type: EMAIL_SENDING
                  status: COMPLETED
                  dependents:
                    - 6ba7b810-9dad-11d1-80b4-00c04fd430c8
                  dependencies: [ ]
                  attemptCount: 1
                  maxAttemptCount: 5
                  result: Email sent to user 123
                  startedAt: '2026-06-25T10:15:30Z'
                  completedAt: '2026-06-25T10:15:32Z'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/jobs/filter:
    get:
      tags: [ Jobs ]
      operationId: filterJobs
      summary: Filter jobs by status and/or type
      description: |
        At least one of `status` or `type` must be provided. When both are given,
        results match both (logical AND). Values are case-insensitive.
      parameters:
        - name: status
          in: query
          required: false
          description: "Filter by status: PENDING, PROCESSING, CANCELED, COMPLETED, FAILED, DEAD."
          schema:
            $ref: '#/components/schemas/JobStatus'
          examples:
            pending:
              summary: Pending jobs
              value: PENDING
            failed:
              summary: Failed jobs
              value: FAILED
            dead:
              summary: Dead jobs
              value: DEAD
        - name: type
          in: query
          required: false
          description: "Filter by type: EMAIL_SENDING, PAYMENT_SENDING."
          schema:
            $ref: '#/components/schemas/JobType'
          examples:
            email:
              summary: Email jobs
              value: EMAIL_SENDING
            payment:
              summary: Payment jobs
              value: PAYMENT_SENDING
        - name: page
          in: query
          required: false
          description: Zero-indexed page number. Defaults to 0.
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Page size. Defaults to 20, must be >= 1.
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        '200':
          description: Filtered jobs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseJobResponseList'
              example:
                message: Operation Successful.
                code: 200
                data:
                  - id: 550e8400-e29b-41d4-a716-446655440000
                    priority: HIGH
                    type: EMAIL_SENDING
                    status: PENDING
                    dependents: [ ]
                    dependencies: [ ]
                    attemptCount: 0
                    maxAttemptCount: 5
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/jobs/{id}/cancel:
    post:
      tags: [ Jobs ]
      operationId: cancelJob
      summary: Cancel a job
      description: |
        Cancels a PENDING or FAILED job. Returns 409 if the job is PROCESSING.
        Terminal states (CANCELED, COMPLETED, DEAD) return 400.
      parameters:
        - $ref: '#/components/parameters/JobIdPath'
      responses:
        '202':
          description: Job cancelled successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                message: Job is cancelled successfully
                code: 202
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/jobs/{id}/revive:
    post:
      tags: [ Jobs ]
      operationId: reviveDeadJobById
      summary: Revive a single dead job
      description: Resets a DEAD job to PENDING with attemptCount=1 and re-publishes to its priority queue.
      parameters:
        - $ref: '#/components/parameters/JobIdPath'
      responses:
        '200':
          description: Job revived successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
              example:
                message: Job revived successfully
                code: 200
                data: 550e8400-e29b-41d4-a716-446655440000
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/jobs/revive:
    post:
      tags: [ Jobs ]
      operationId: reviveAllDeadJobs
      summary: Revive all dead jobs
      description: |
        Finds all DEAD jobs, resets each to PENDING with attemptCount=1, and
        re-publishes to their original priority queues. Returns 404 if no
        dead jobs exist.
      responses:
        '200':
          description: Dead jobs revived successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseStringList'
              example:
                message: Operation Successful.
                code: 200
                data:
                  - 550e8400-e29b-41d4-a716-446655440000
                  - 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/dashboard/jobs/summary:
    get:
      tags: [ Dashboard ]
      operationId: getJobsSummary
      summary: Job counts by status
      description: Returns counts for each status (PENDING, PROCESSING, CANCELED, COMPLETED, FAILED, DEAD) plus a TOTAL. Read-only, suitable for frequent polling.
      responses:
        '200':
          description: Job status summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseJobsSummary'
              example:
                message: Operation Successful.
                code: 200
                data:
                  PENDING: 12
                  PROCESSING: 3
                  CANCELED: 1
                  COMPLETED: 45
                  FAILED: 2
                  DEAD: 1
                  TOTAL: 64
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/dashboard/queue-metrics:
    get:
      tags: [ Dashboard ]
      operationId: getQueueMetrics
      summary: Queue message counts
      description: |
        Current message count for each RabbitMQ queue: HIGH/MEDIUM/LOW priority,
        RETRY (30s TTL), and DEAD_LETTER. All queues are quorum queues with
        reject-publish overflow.
      responses:
        '200':
          description: Queue metrics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseQueueMetrics'
              example:
                message: Operation Successful.
                code: 200
                data:
                  queues:
                    HIGH_PRIORITY_QUEUE_LENGTH: 4
                    MEDIUM_PRIORITY_QUEUE_LENGTH: 7
                    LOW_PRIORITY_QUEUE_LENGTH: 15
                    RETRY_QUEUE_LENGTH: 2
                    DEAD_LETTER_QUEUE_LENGTH: 1
        '500':
          $ref: '#/components/responses/InternalError'

components:
  parameters:
    JobIdPath:
      name: id
      in: path
      required: true
      description: UUID v4 of the target job.
      schema:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000

  responses:
    BadRequest:
      description: Bad request — validation error, malformed JSON, or business rule violation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CommonResponseError'
          examples:
            validation:
              summary: Field-level validation error
              value:
                message: "Validation error: payload: Payload cannot be blank"
                code: 400
            malformedJson:
              summary: Malformed JSON
              value:
                message: Malformed JSON request
                code: 400
            businessRule:
              summary: Business rule violation
              value:
                message: At least one of 'status' or 'type' must be provided
                code: 400
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CommonResponseError'
          examples:
            jobNotFound:
              summary: Job not found
              value:
                message: No job found with given id.
                code: 404
            noDeadJobs:
              summary: No dead jobs to revive
              value:
                message: No dead jobs found to revive
                code: 404
    Conflict:
      description: Conflict — job is currently PROCESSING and cannot be cancelled.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CommonResponseError'
          example:
            message: Conflict
            code: 409
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CommonResponseError'
          example:
            message: Internal server error
            code: 500

  schemas:
    JobPriority:
      title: JobPriority
      type: string
      description: "Priority level. HIGH > MEDIUM > LOW. Enforced for dependencies."
      enum: [ HIGH, MEDIUM, LOW ]
      example: HIGH

    JobType:
      title: JobType
      type: string
      description: "Job type. Each type maps to a specific worker handler."
      enum: [ EMAIL_SENDING, PAYMENT_SENDING ]
      example: EMAIL_SENDING

    JobStatus:
      title: JobStatus
      type: string
      description: |
        Lifecycle status. State machine:
        `PENDING -> PROCESSING -> COMPLETED | FAILED -> PENDING (retry) -> DEAD | CANCELED`
      enum: [ PENDING, PROCESSING, CANCELED, COMPLETED, FAILED, DEAD ]
      example: PENDING

    JobRequest:
      title: JobRequest
      type: object
      description: |
        Create job request. `dependencies` and `maxAttemptCount` are optional.
        Dependencies are validated for existence, non-terminal status, and priority hierarchy.
        A job starts with status=PENDING and attemptCount=1.
      required: [ priority, type, payload ]
      properties:
        priority:
          $ref: '#/components/schemas/JobPriority'
        type:
          $ref: '#/components/schemas/JobType'
        payload:
          title: Payload
          type: string
          description: Job payload. Non-blank, max 10,000 characters.
          minLength: 1
          maxLength: 10000
          example: Send welcome email to user 123
        dependencies:
          title: Dependencies
          type: array
          description: UUIDs of jobs that must complete first. No upper limit. Defaults to [].
          items:
            type: string
            format: uuid
          example:
            - 550e8400-e29b-41d4-a716-446655440000
            - 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        maxAttemptCount:
          title: MaxAttemptCount
          type: integer
          description: Max retries before DEAD. Default 5, min 1, no upper limit.
          minimum: 1
          nullable: true
          example: 3

    JobResponse:
      title: JobResponse
      type: object
      description: Job representation. Null fields are omitted (NON_NULL serialization).
      properties:
        id:
          title: ID
          type: string
          format: uuid
          description: Unique identifier (UUID v4).
        priority:
          $ref: '#/components/schemas/JobPriority'
        type:
          $ref: '#/components/schemas/JobType'
        status:
          $ref: '#/components/schemas/JobStatus'
        dependents:
          title: Dependents
          type: array
          description: UUIDs of jobs that depend on this job.
          items:
            type: string
            format: uuid
        dependencies:
          title: Dependencies
          type: array
          description: UUIDs of jobs this job depends on.
          items:
            type: string
            format: uuid
        result:
          title: Result
          type: string
          description: Output from the worker. Only for COMPLETED jobs.
          nullable: true
        attemptCount:
          title: Attempt Count
          type: integer
          description: Number of processing attempts so far.
          nullable: true
        maxAttemptCount:
          title: Max Attempt Count
          type: integer
          description: Maximum attempts before DEAD.
          nullable: true
        startedAt:
          title: Started At
          type: string
          format: date-time
          description: When processing started (ISO-8601).
          nullable: true
        completedAt:
          title: Completed At
          type: string
          format: date-time
          description: When processing completed (ISO-8601).
          nullable: true

    CommonResponse:
      title: CommonResponse
      type: object
      description: Standard response envelope. On errors, `data` is omitted.
      properties:
        message:
          title: Message
          type: string
          description: Human-readable status message.
          default: Operation Successful.
        code:
          title: Code
          type: integer
          description: HTTP status code mirrored in the body.
          default: 200
        data:
          title: Data
          description: Response payload. Type varies by endpoint. Omitted on errors.
        traceId:
          title: Trace ID
          type: string
          description: Distributed trace identifier (nullable).
          nullable: true
        spanId:
          title: Span ID
          type: string
          description: Span identifier within the trace (nullable).
          nullable: true

    CommonResponseError:
      title: CommonResponseError
      description: Error envelope. `data` is always omitted.
      allOf:
        - $ref: '#/components/schemas/CommonResponse'

    CommonResponseJobResponse:
      title: CommonResponseJobResponse
      description: CommonResponse with `data` as a single JobResponse.
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/JobResponse'

    CommonResponseJobResponseList:
      title: CommonResponseJobResponseList
      description: CommonResponse with `data` as an array of JobResponse.
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/JobResponse'

    CommonResponseString:
      title: CommonResponseString
      description: CommonResponse with `data` as a string (e.g. revived job UUID).
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: string

    CommonResponseStringList:
      title: CommonResponseStringList
      description: CommonResponse with `data` as an array of strings (e.g. revived job UUIDs).
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                type: string

    JobsSummary:
      title: JobsSummary
      type: object
      description: Job counts per status plus TOTAL.
      properties:
        PENDING:
          type: integer
        PROCESSING:
          type: integer
        CANCELED:
          type: integer
        COMPLETED:
          type: integer
        FAILED:
          type: integer
        DEAD:
          type: integer
        TOTAL:
          type: integer

    CommonResponseJobsSummary:
      title: CommonResponseJobsSummary
      description: CommonResponse with `data` as a JobsSummary.
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/JobsSummary'

    QueueMetrics:
      title: QueueMetrics
      type: object
      description: Message counts for each RabbitMQ queue.
      properties:
        queues:
          title: Queues
          type: object
          properties:
            HIGH_PRIORITY_QUEUE_LENGTH:
              type: integer
            MEDIUM_PRIORITY_QUEUE_LENGTH:
              type: integer
            LOW_PRIORITY_QUEUE_LENGTH:
              type: integer
            RETRY_QUEUE_LENGTH:
              type: integer
            DEAD_LETTER_QUEUE_LENGTH:
              type: integer

    CommonResponseQueueMetrics:
      title: CommonResponseQueueMetrics
      description: CommonResponse with `data` as QueueMetrics.
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/QueueMetrics'