# producer/src/main/java/com/nadimnesar/jobqueue/producer/dto/request/JobRequest.java
package com.nadimnesar.jobqueue.producer.dto.request;

import com.nadimnesar.jobqueue.common.constant.enums.JobPriority;
import com.nadimnesar.jobqueue.common.constant.enums.JobType;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.HashSet;
import java.util.Set;
import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class JobRequest {

    @NotNull(message = "Job priority cannot be null")
    private JobPriority priority;

    @NotNull(message = "Job type cannot be null")
    private JobType type;

    @Builder.Default
    private Set<UUID> dependencies = new HashSet<>();

    @NotBlank(message = "Payload cannot be blank")
    @Size(max = 10000, message = "Payload cannot exceed 10000 characters")
    private String payload;

    @Min(value = 1, message = "Max attempt count must be at least 1")
    private Integer maxAttemptCount;
}